+++ src/bika/lims/content/analysisrequest.py
@@ -58,7 +58,6 @@ from bika.lims.interfaces import ISubmitted
 from bika.lims.utils import getUsers
 from bika.lims.utils import tmpID
 from bika.lims.utils import user_email
-from bika.lims.utils import user_fullname
 from bika.lims.utils.analysisrequest import apply_hidden_services
 from bika.lims.workflow import getTransitionDate
 from bika.lims.workflow import getTransitionUsers
@@ -1150,12 +1149,6 @@ schema = BikaSchema.copy() + Schema((
         expression='here.getReceivedBy()',
         default='',
         widget=ComputedWidget(visible=False,),
-    ),
-
-    ComputedField(
-        'CreatorFullName',
-        expression="here._getCreatorFullName()",
-        widget=ComputedWidget(visible=False),
     ),
 
     ComputedField(
@@ -2255,12 +2248,6 @@ class AnalysisRequest(BaseFolder, ClientAwareMixin):
             analysis_obj = analysis.getObject()
             catalog.reindexObject(analysis_obj, idxs=idxs, update_metadata=1)
 
-    def _getCreatorFullName(self):
-        """
-        Returns the full name of this analysis request's creator.
-        """
-        return user_fullname(self, self.Creator())
-
     def _getCreatorEmail(self):
         """
         Returns the email of this analysis request's creator.
@@ -2271,7 +2258,7 @@ class AnalysisRequest(BaseFolder, ClientAwareMixin):
         """
         Returns the full name's defined sampler.
         """
-        return user_fullname(self, self.getSampler())
+        return api.get_user_fullname(self.getSampler())
 
     def _getSamplerEmail(self):
         """