+++ src/senaite/core/catalog/report_catalog.py
@@ -39,7 +39,6 @@ COLUMNS = BASE_COLUMNS + [
     # attribute name
     "getClientTitle",
     "getClientURL",
-    "getCreatorFullName",
     "getFileSize",
 ]